Prevent tabbers from displaying tooltips on _ALL_ tabber content
This is something that basically always annoyed me, but I never really
took the time to do something about it.
Tabbers have a title="" attribute for the tab content of all tabbers,
which doesn't add anything to the functionality other than annoyingly
display a tooltip showing which tab you're currently on. Just like how
I would not like to have a <body title="{{PAGENAME}}"> or something,
this is just something that doesn't add anything, and if anything, it's
annoying.

I did not change the fact the anchors at the top of the tab still get a
title="" attribute, since it makes sense to use title-attributes there
(just like how wikilinks get a title assigned to them).

Data-attributes were designed exactly for this purpose, so let's
actually use them to switch between tabs.

I've tested this locally in my browser, but feel free to test the
functionality further.